Seperate useTableSession hook for session time management
